Extracted model class to its own file
